backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
Markdown==3.4.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
riotwatcher==3.2.3
sqlparse==0.4.2
urllib3==1.26.11
